1.更改数据库金额为string类型并修改后端代码 2.使用bigdecimal进行计算，增加bigdecimal计算工具类 3.修改cotroller接收参数方式

+++ src/main/java/priv/hsy/redenvelops/entity/RedDetail.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.*;
 
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 @Getter
@@ -13,9 +14,9 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 public class RedDetail {
     @TableId(type = IdType.AUTO)//主键自增 数据库中需要设置主键自增
-    private Integer id;
-    private Integer rid;
-    private Integer receiveId;
+    private BigInteger id;
+    private BigInteger rid;
+    private BigInteger receiveId;
     private Double getMoney;
     private Timestamp getTime;
